Support building Win32 DLLs without relying on auto-import
This allows to build the DLLs with other (e.g. Microsoft) compilers or
when --disable-auto-import is explicitly specified with GNU ld.

Define XMLWRAPP_API and XSLTWRAPP_API as appropriate and add missing
occurrences of these macros and also remove a couple of superfluous
ones, as this resulted in -Wattributes from gcc.